Split TableStructureValidator
