+++ libs/output-mapping/src/Exception/TableNotFoundException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Keboola\OutputMapping\Exception;
+
+use RuntimeException;
+
+class TableNotFoundException extends RuntimeException
+{
+
+}
